Add into GateLinear
